Add files via upload
Thass a sample to my charity website, where I've created a space to help make inspiring stories. The Donation Wall is a special feature that highlights the impact of donors and acknowledges their contributions. On the Stories page, you'll find touching accounts of the lives changed by the support received. I've also included a Contact Us page for easy communication. The entire site is designed with a beautiful, calming background to ensure a pleasant and inspiring experience.